float arg 2

+++ training_scripts/train_immunogenicity_model.py
@@ -683,7 +683,7 @@ if __name__=="__main__":
     parser.add_argument("--LDS_ks", type=int, default=11)
     parser.add_argument("--LDS_sigma", type=float, default=1.0)
     parser.add_argument("--loss_weights", type=str, default="LDS_weights")
-    parser.add_argument("--constraints_weight", type=int, default=0.00)
+    parser.add_argument("--constraints_weight", type=float, default=0.00)
     
     parser.add_argument("--lr", type=float, default=3e-4)
     parser.add_argument("--weight_decay", type=float, default=0.0001)
